Convert the SubmitCircomProof component into a function
